requirements.txt
astroid==2.15.2
blinker==1.6.1
click==8.1.3
dill==0.3.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-Security-Too==5.1.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mypy==1.2.0
mypy-extensions==1.0.0
passlib==1.7.4
platformdirs==3.2.0
psycopg2-binary==2.9.6
pylint==2.17.2
SQLAlchemy==2.0.9
tomlkit==0.11.7
typing_extensions==4.5.0
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
